requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Gunicorn==21.2.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
squareup==36.1.0
